relocate some snapshot files
